requirements.txt
Django>=5.0.2
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.0